Domain driven development (#2)
* Update department views (frontend & backend). Add notifications to json requests

* Refactor Project to Domain Driven Design.

The project was initially written using the repository design pattern without
the need for a scale . After a thorough analysis of the countless
features (HRM, HelpDesk, Biometric integration ... ) required, we opted
to use the domain driven design pattern to handle such requirement. This
has the advantage of decoupling the application into components that
can be tested, added or removed individually without breaking the application.